restructure the route into the controller folder according to mvc practices
